requirements.txt
Pillow==9.2.0
tqdm==4.64.1
